llm_automl_project/backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
numpy
requests
python-multipart
aiofiles
fastapi
uvicorn
pandas
scikit-learn
sqlalchemy
fastapi
uvicorn
pandas
scikit-learn
sqlalchemy
python-dotenv
